+++ guillotina_elasticsearch/__init__.py
@@ -17,6 +17,7 @@ app_settings = {
     },
     'commands': {
         'es-migrate': 'guillotina_elasticsearch.commands.migrate.MigrateCommand',
+        'es-reindex': 'guillotina_elasticsearch.commands.reindex.ReindexCommand',
         'es-vacuum': 'guillotina_elasticsearch.commands.vacuum.VacuumCommand'
     }
 }